refactor(consensus): rename state.rs to epoch_state.rs
